+++ CHANGELOG.md
@@ -6,12 +6,16 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a CH
 
 ### Added
 - Add back button to test source editor.
+- Add details page with compilation results in the test source list.
 
 ### Changed
 - Fix style at live machine list.
 - Fix state after saving test source file.
 - Changed naming monkey runner => monkey tool.
 - Fix style of status boxes in live mode.
+
+### Removed
+- Removed from APK Manager list the apks that come from test sources.
 
 ## [0.7.2]